+++ src/components/Search/Search.js
@@ -1,6 +1,10 @@
 import { React, useEffect, useState, useRef } from 'react'
 import './Search.css'
-import { envTilerURL, constructTilerParams } from './envVarSetup'
+import {
+  envTilerURL,
+  constructTilerParams,
+  envMosaicTilerURL
+} from './envVarSetup'
 import { convertDateTimeForAPI, debounce } from '../../utils'
 import { MIN_ZOOM } from '../defaults'
 
@@ -37,6 +41,7 @@ const Search = () => {
     (state) => state.mainSlice.currentPopupResult
   )
   const tilerURL = envTilerURL
+  const mosaicTilerURL = envMosaicTilerURL
 
   // set up map state
   const map = _map
@@ -46,17 +51,18 @@ const Search = () => {
 
   // set up local state
   const [dateTimeValue, setDateTimeValue] = useState([twoWeeksAgo, new Date()])
-  const [resultFootprints, setResultFootprints] = useState()
-  const [clickedFootprintHighlights, setClickedFootprintsHighlight] = useState()
-  const [clickedFootprintsImageLayer, setClickedFootprintsImageLayer] =
-    useState()
   const [collectionError, setCollectionError] = useState(false)
   const [zoomLevelValue, setZoomLevelValue] = useState(0)
+
   const dateTimeRef = useRef(dateTimeValue)
   const zoomLevelRef = useRef(0)
   const selectedCollectionRef = useRef(_collectionSelected)
   const showCloudSliderRef = useRef(_showCloudSlider)
   const viewModeRef = useRef(_viewMode)
+
+  const resultFootprintsRef = useRef()
+  const clickedFootprintHighlightRef = useRef()
+  const clickedFootprintImageLayerRef = useRef()
 
   // override leaflet draw tooltips
   // eslint-disable-next-line no-import-assign
@@ -90,20 +96,21 @@ const Search = () => {
 
       // set up layerGroup for footprints and add to map
       const resultFootprintsInit = new L.FeatureGroup()
-      setResultFootprints(resultFootprintsInit)
       resultFootprintsInit.addTo(map)
       resultFootprintsInit.id = 'resultFootprints'
+      resultFootprintsRef.current = resultFootprintsInit
 
       // set up layerGroup for highlight footprints and add to map
       const clickedFootprintsHighlightInit = new L.FeatureGroup()
-      setClickedFootprintsHighlight(clickedFootprintsHighlightInit)
       clickedFootprintsHighlightInit.addTo(map)
-      clickedFootprintsHighlightInit.id = 'clickedFootprintHighlights'
+      clickedFootprintsHighlightInit.id = 'clickedFootprintHighlight'
+      clickedFootprintHighlightRef.current = clickedFootprintsHighlightInit
 
       // set up layerGroup for image layer and add to map
-      const clickedFootprintsImageLayerInit = new L.FeatureGroup()
-      setClickedFootprintsImageLayer(clickedFootprintsImageLayerInit)
-      clickedFootprintsImageLayerInit.addTo(map)
+      const clickedFootprintImageLayerInit = new L.FeatureGroup()
+      clickedFootprintImageLayerInit.addTo(map)
+      clickedFootprintImageLayerInit.id = 'clickedFootprintImageLayer'
+      clickedFootprintImageLayerRef.current = clickedFootprintImageLayerInit
 
       // setup custom pane for tiler image result
       map.createPane('imagery')
@@ -122,24 +129,24 @@ const Search = () => {
       map.on('zoomend', function () {
         zoomLevelRef.current = map.getZoom()
         setZoomLevelValue(map.getZoom())
-        processSearch(resultFootprintsInit)
+        processSearch()
       })
 
       map.on('dragend', function () {
-        processSearch(resultFootprintsInit)
+        processSearch()
       })
 
       map.on('click', mapClickHandler)
 
       // initiate first search on load
-      processSearch(resultFootprintsInit)
+      processSearch()
     }
   }, [map])
 
   // when zoom level changes, set in global store to hide/show zoom notice
   // and perform search if within zoom limits
   useEffect(() => {
-    if (zoomLevelValue >= MIN_ZOOM || _viewMode === 'footprint') {
+    if (zoomLevelValue >= MIN_ZOOM || _viewMode === 'scene') {
       dispatch(setShowZoomNotice(false))
     } else {
       dispatch(setShowZoomNotice(true))
@@ -172,10 +179,9 @@ const Search = () => {
   // when currentPopupResult set, add image layer to map
   useEffect(() => {
     if (_currentPopupResult !== null) {
-      if (clickedFootprintsImageLayer) {
-        clickedFootprintsImageLayer.clearLayers()
-      }
-      // call add new image layer to map function
+      clickedFootprintImageLayerRef.current.clearLayers()
+
+      // call add new tiler image layer to map function
       addImageClicked(_currentPopupResult)
     }
   }, [_currentPopupResult])
@@ -185,15 +191,13 @@ const Search = () => {
   // the user clicks just on the map
   function mapClickHandler(e) {
     // if double-clicking the image, zoom in, otherwise process click
-    if (e.originalEvent.detail === 2) return
+    // disable click function in mosaic view
+    if (e.originalEvent.detail === 2 || viewModeRef.current === 'mosaic') return
+
     const clickBounds = L.latLngBounds(e.latlng, e.latlng)
 
-    if (clickedFootprintHighlights) {
-      clickedFootprintHighlights.clearLayers()
-    }
-    if (clickedFootprintsImageLayer) {
-      clickedFootprintsImageLayer.clearLayers()
-    }
+    clickedFootprintHighlightRef.current.clearLayers()
+    clickedFootprintImageLayerRef.current.clearLayers()
 
     const intersectingFeatures = []
 
@@ -201,9 +205,9 @@ const Search = () => {
       for (const f in _searchResults.features) {
         const feature = _searchResults.features[f]
         const bounds = feature.bbox
-        const latlng1 = L.latLng(bounds[1], bounds[0])
-        const latlng2 = L.latLng(bounds[3], bounds[2])
-        const featureBounds = L.latLngBounds(latlng1, latlng2)
+        const swCorner = L.latLng(bounds[1], bounds[0])
+        const neCorner = L.latLng(bounds[3], bounds[2])
+        const featureBounds = L.latLngBounds(swCorner, neCorner)
         if (featureBounds && clickBounds.intersects(featureBounds)) {
           intersectingFeatures.push(feature)
           // add features to clickedHighlight layer
@@ -216,7 +220,7 @@ const Search = () => {
           const clickedFootprintsFound = L.geoJSON(feature, {
             style: clickedFootprintsSelectedStyle
           })
-          clickedFootprintsFound.addTo(clickedFootprintHighlights)
+          clickedFootprintsFound.addTo(clickedFootprintHighlightRef.current)
         }
       }
     }
@@ -231,7 +235,7 @@ const Search = () => {
     }
   }
 
-  function clearResults(resultFootPrintsLocal) {
+  function clearResults() {
     // show loading spinner, clear previous results
     batch(() => {
       dispatch(setSearchResults(null))
@@ -239,33 +243,26 @@ const Search = () => {
     })
 
     // remove clicked footprint highlight
-    if (clickedFootprintHighlights) clickedFootprintHighlights.clearLayers()
+    if (clickedFootprintHighlightRef)
+      clickedFootprintHighlightRef.current.clearLayers()
 
     // remove image layer
-    if (clickedFootprintsImageLayer) clickedFootprintsImageLayer.clearLayers()
+    if (clickedFootprintImageLayerRef)
+      clickedFootprintImageLayerRef.current.clearLayers()
 
     // remove existing footprints from map
-    removeFootprints(resultFootPrintsLocal)
-  }
-
-  // remove old footprints from map
-  function removeFootprints(resultFootPrintsLocal) {
-    resultFootPrintsLocal.eachLayer(function (layer) {
-      resultFootPrintsLocal.removeLayer(layer)
+    resultFootprintsRef.current.eachLayer(function (layer) {
+      resultFootprintsRef.current.removeLayer(layer)
     })
   }
 
   // search throttle set to 1500ms
-  const processSearch = (resultFootprintsInit) =>
-    debounce(searchAPI(resultFootprintsInit), 1500)
+  const processSearch = () => debounce(searchAPI(), 1500)
 
   // function called when search is initiated
-  function searchAPI(resultFootprintsInit) {
-    // set up footprints layer
-    const resultFootPrintsLocal = resultFootprintsInit || resultFootprints
-
+  function searchAPI() {
     // clear previous results from map
-    clearResults(resultFootPrintsLocal)
+    clearResults()
 
     // if the zoom level is too high in mosaic view, abort search
     // otherwise, move to the mosaic search
@@ -275,7 +272,7 @@ const Search = () => {
       zoomLevelRef.current >= MIN_ZOOM &&
       viewModeRef.current === 'mosaic'
     ) {
-      addMosaic(resultFootPrintsLocal)
+      addMosaic()
       return
     }
 
@@ -306,7 +303,6 @@ const Search = () => {
       viewportBounds._northEast.lng,
       viewportBounds._northEast.lat
     ].join(',')
-
     let cloudCover = ''
     if (showCloudSliderRef.current)
       cloudCover = `query=%7B"eo%3Acloud_cover"%3A%7B"gte"%3A0,"lte"%3A${_cloudCover}%7D%7D`
@@ -341,7 +337,7 @@ const Search = () => {
 
         // add new footprints to map
         const resultFootprintsFound = L.geoJSON(json, {})
-        resultFootprintsFound.addTo(resultFootPrintsLocal)
+        resultFootprintsFound.addTo(resultFootprintsRef.current)
       })
   }
 
@@ -350,7 +346,7 @@ const Search = () => {
     // show loading spinner
     dispatch(setSearchLoading(true))
 
-    clickedFootprintsImageLayer.clearLayers()
+    clickedFootprintImageLayerRef.current.clearLayers()
     const featureURL = feature.links[0].href
     const tilerParams = constructTilerParams(_collectionSelected)
 
@@ -361,9 +357,9 @@ const Search = () => {
         return response.json()
       })
       .then(function (json) {
-        const corner1 = L.latLng(json.bbox[1], json.bbox[0])
-        const corner2 = L.latLng(json.bbox[3], json.bbox[2])
-        const tileBounds = L.latLngBounds(corner1, corner2)
+        const swCorner = L.latLng(json.bbox[1], json.bbox[0])
+        const neCorner = L.latLng(json.bbox[3], json.bbox[2])
+        const tileBounds = L.latLngBounds(swCorner, neCorner)
 
         L.tileLayer(
           `${tilerURL}/stac/tiles/{z}/{x}/{y}.png?url=${featureURL}${tilerParams}`,
@@ -374,7 +370,7 @@ const Search = () => {
             pane: 'imagery'
           }
         )
-          .addTo(clickedFootprintsImageLayer)
+          .addTo(clickedFootprintImageLayerRef.current)
           .on('load', function () {
             // hide loading spinner
             dispatch(setSearchLoading(false))
@@ -386,9 +382,12 @@ const Search = () => {
   }
 
   // function to remove old image layer and add new Tiler image layer to map
-  function addMosaic(resultFootPrintsLocal) {
+  function addMosaic() {
     // clear previous results from map
-    clearResults(resultFootPrintsLocal)
+    clearResults()
+
+    // show loading spinner
+    dispatch(setSearchLoading(true))
 
     // const tilerParams = constructTilerParams(_collectionSelected)
 
@@ -407,9 +406,9 @@ const Search = () => {
       viewportBounds._northEast.lat
     ]
 
-    const corner1 = L.latLng(bbox[1], bbox[0])
-    const corner2 = L.latLng(bbox[3], bbox[2])
-    const mosaicBounds = L.latLngBounds(corner1, corner2)
+    const swCorner = L.latLng(bbox[1], bbox[0])
+    const neCorner = L.latLng(bbox[3], bbox[2])
+    const mosaicBounds = L.latLngBounds(swCorner, neCorner)
 
     const requestOptions = {
       method: 'POST',
@@ -421,22 +420,22 @@ const Search = () => {
         asset_name: 'visual', // todo: use first entry in assets config
         collections: [selectedCollectionRef.current],
         datetime,
-        bbox
+        bbox,
+        max_items: 10
         // todo: query w/ cloud_cover, as a JSON object rather than a url-encoded string
       })
     }
-    fetch(`${tilerURL}/mosaicjson/mosaics`, requestOptions)
+    fetch(`${mosaicTilerURL}/mosaicjson/mosaics`, requestOptions)
       .then((r) => r.json())
       .then((body) => {
-        console.log(`mosaic create response: ${JSON.stringify(body)}`)
-        const tileHref = body.links[3].href // todo: rel=tiles
+        const tileHref = body?.links?.find((el) => el.rel === 'tiles').href
         L.tileLayer(`${tileHref}`, {
           attribution: '©OpenStreetMap',
           tileSize: 256,
           bounds: mosaicBounds,
           pane: 'imagery'
         })
-          .addTo(clickedFootprintsImageLayer)
+          .addTo(clickedFootprintImageLayerRef.current)
           .on('load', function () {
             // hide loading spinner
             dispatch(setSearchLoading(false))
@@ -456,7 +455,7 @@ const Search = () => {
       </div>
       <div className="searchContainer datePicker">
         <label>
-          Select Date Range{' '}
+          Date Range{' '}
           {!dateTimeValue && (
             <span className="error-true">
               <em>Required</em>
@@ -475,9 +474,11 @@ const Search = () => {
       <div className="searchContainer cloudSlider">
         <CloudSlider></CloudSlider>
       </div>
-      <div className="searchContainer">
-        <ViewSelector></ViewSelector>
-      </div>
+      {mosaicTilerURL && (
+        <div className="searchContainer">
+          <ViewSelector></ViewSelector>
+        </div>
+      )}
     </div>
   )
 }